requirements.txt
torch
torchvision
pillow
numpy
opencv-python
einops
timm
pyyaml
